public/data/repos/kaggle_mcp/content/requirements.txt
fastmcp
kaggle